Supplementary Unit Test Cases.
